- Refactor package to help make navigating to various utils and helpers easier
