+++ external/user/v2/utils.go
@@ -89,81 +89,6 @@ func (s *DefaultStringUtils) InSlice(item string, slice []string) bool {
 	return false
 }
 
-// UserFactory provides convenient user creation
-type UserFactory struct {
-	config       *UserConfig
-	idGenerator  IDGenerator
-	timeProvider TimeProvider
-	stringUtils  StringUtils
-}
-
-// NewUserFactory creates a new user factory with default implementations
-func NewUserFactory(config *UserConfig) *UserFactory {
-	if config == nil {
-		config = DefaultUserConfig()
-	}
-
-	return &UserFactory{
-		config:       config,
-		idGenerator:  &DefaultIDGenerator{},
-		timeProvider: &DefaultTimeProvider{},
-		stringUtils:  &DefaultStringUtils{},
-	}
-}
-
-// NewUserFactoryWithDependencies creates a factory with custom implementations
-func NewUserFactoryWithDependencies(
-	config *UserConfig,
-	idGenerator IDGenerator,
-	timeProvider TimeProvider,
-	stringUtils StringUtils,
-) *UserFactory {
-	if config == nil {
-		config = DefaultUserConfig()
-	}
-
-	return &UserFactory{
-		config:       config,
-		idGenerator:  idGenerator,
-		timeProvider: timeProvider,
-		stringUtils:  stringUtils,
-	}
-}
-
-// CreateUser creates a new user with initial setup
-func (f *UserFactory) CreateUser(email string) *UniversalUser {
-	user := NewUniversalUser(f.config, f.idGenerator, f.timeProvider, f.stringUtils)
-
-	user.Email = f.stringUtils.ToLowerCase(email)
-	user.GenerateNewUUID()
-
-	if f.config.MultipleIdentifiers {
-		user.GenerateNewNanoID()
-	}
-
-	user.SetInitialState()
-
-	return user
-}
-
-// CreateUserWithPersonalInfo creates a user with personal information
-func (f *UserFactory) CreateUserWithPersonalInfo(email, firstName, lastName string) *UniversalUser {
-	user := f.CreateUser(email)
-
-	user.PersonalInfo.FirstName = f.stringUtils.ToTitleCase(firstName)
-	user.PersonalInfo.LastName = f.stringUtils.ToTitleCase(lastName)
-	user.PersonalInfo.FullName = fmt.Sprintf("%s %s",
-		user.PersonalInfo.FirstName,
-		user.PersonalInfo.LastName)
-
-	return user
-}
-
-// LoadExistingUser loads an existing user and sets up dependencies
-func (f *UserFactory) LoadExistingUser(user *UniversalUser) *UniversalUser {
-	return user.SetDependencies(f.config, f.idGenerator, f.timeProvider, f.stringUtils)
-}
-
 // Migration helpers for existing projects
 
 // MigrateFromLegacyUser converts your existing User to UniversalUser
@@ -235,55 +160,4 @@ func MigrateToLegacyUser(universalUser *UniversalUser) *user.User {
 	}
 
 	return legacyUser
-}
-
-// Project-specific configurations
-
-// WebAppUserConfig returns configuration suitable for web applications
-func WebAppUserConfig() *UserConfig {
-	return &UserConfig{
-		DefaultStatus: "PROVISIONED",
-		StatusTransitions: map[string][]string{
-			"ACTIVE":      {"PROVISIONED"},
-			"SUSPENDED":   {"ACTIVE"},
-			"DEACTIVATED": {"ACTIVE", "SUSPENDED"},
-			"REACTIVATE":  {"DEACTIVATED"},
-			"UNSUSPEND":   {"SUSPENDED"},
-		},
-		RequiredFields:            []string{"email", "first_name", "last_name"},
-		ValidRoles:                []string{"ADMIN", "USER", "MODERATOR"},
-		EmailVerificationRequired: true,
-		MultipleIdentifiers:       false,
-	}
-}
-
-// APIServiceUserConfig returns configuration suitable for API services
-func APIServiceUserConfig() *UserConfig {
-	return &UserConfig{
-		DefaultStatus: "ACTIVE",
-		StatusTransitions: map[string][]string{
-			"ACTIVE":    {"PROVISIONED"},
-			"SUSPENDED": {"ACTIVE"},
-			"DISABLED":  {"ACTIVE", "SUSPENDED"},
-		},
-		RequiredFields:            []string{"email"},
-		ValidRoles:                []string{"SERVICE", "CLIENT", "ADMIN"},
-		EmailVerificationRequired: false,
-		MultipleIdentifiers:       true,
-	}
-}
-
-// MicroserviceUserConfig returns minimal configuration for microservices
-func MicroserviceUserConfig() *UserConfig {
-	return &UserConfig{
-		DefaultStatus: "ACTIVE",
-		StatusTransitions: map[string][]string{
-			"ACTIVE":   {},
-			"INACTIVE": {"ACTIVE"},
-		},
-		RequiredFields:            []string{"email"},
-		ValidRoles:                []string{}, // Allow any roles
-		EmailVerificationRequired: false,
-		MultipleIdentifiers:       true,
-	}
 }
